Prepare criteria edit page layout for judges list
